add more error parameter checks

+++ fs.js
@@ -2,36 +2,33 @@
 // Use of this source code is governed by a MIT-style license that can be
 // found in the LICENSE file.
 
-import {
-  NativeModules,
-  DeviceEventEmitter,
-  Platform,
-  NativeAppEventEmitter,
-} from 'react-native'
+// import type {RNFetchBlobConfig, RNFetchBlobNative, RNFetchBlobStream} from './types'
+
+import {NativeModules, Platform} from 'react-native'
 import RNFetchBlobSession from './class/RNFetchBlobSession'
 import RNFetchBlobWriteStream from './class/RNFetchBlobWriteStream'
 import RNFetchBlobReadStream from './class/RNFetchBlobReadStream'
 import RNFetchBlobFile from './class/RNFetchBlobFile'
-import type {
-  RNFetchBlobNative,
-  RNFetchBlobConfig,
-  RNFetchBlobStream
-} from './types'
 
-const RNFetchBlob:RNFetchBlobNative = NativeModules.RNFetchBlob
+const RNFetchBlob: RNFetchBlobNative = NativeModules.RNFetchBlob
 
 const dirs = {
-    DocumentDir :  RNFetchBlob.DocumentDir,
-    CacheDir : RNFetchBlob.CacheDir,
-    PictureDir : RNFetchBlob.PictureDir,
-    MusicDir : RNFetchBlob.MusicDir,
-    MovieDir : RNFetchBlob.MovieDir,
-    DownloadDir : RNFetchBlob.DownloadDir,
-    DCIMDir : RNFetchBlob.DCIMDir,
-    SDCardDir : RNFetchBlob.SDCardDir,
-    SDCardApplicationDir : RNFetchBlob.SDCardApplicationDir,
-    MainBundleDir : RNFetchBlob.MainBundleDir,
-    LibraryDir : RNFetchBlob.LibraryDir
+  DocumentDir: RNFetchBlob.DocumentDir,
+  CacheDir: RNFetchBlob.CacheDir,
+  PictureDir: RNFetchBlob.PictureDir,
+  MusicDir: RNFetchBlob.MusicDir,
+  MovieDir: RNFetchBlob.MovieDir,
+  DownloadDir: RNFetchBlob.DownloadDir,
+  DCIMDir: RNFetchBlob.DCIMDir,
+  SDCardDir: RNFetchBlob.SDCardDir,
+  SDCardApplicationDir: RNFetchBlob.SDCardApplicationDir,
+  MainBundleDir: RNFetchBlob.MainBundleDir,
+  LibraryDir: RNFetchBlob.LibraryDir
+}
+
+function addCode(code: string, error: Error): Error {
+  error.code = code
+  return error
 }
 
 /**
@@ -39,9 +36,9 @@ const dirs = {
  * @param  {string} name Stream ID
  * @return {RNFetchBlobSession}
  */
-function session(name:string):RNFetchBlobSession {
+function session(name: string): RNFetchBlobSession {
   let s = RNFetchBlobSession.getSession(name)
-  if(s)
+  if (s)
     return new RNFetchBlobSession(name)
   else {
     RNFetchBlobSession.setSession(name, [])
@@ -49,25 +46,20 @@ function session(name:string):RNFetchBlobSession {
   }
 }
 
-function asset(path:string):string {
-  if(Platform.OS === 'ios') {
+function asset(path: string): string {
+  if (Platform.OS === 'ios') {
     // path from camera roll
-    if(/^assets-library\:\/\//.test(path))
+    if (/^assets-library\:\/\//.test(path))
       return path
   }
   return 'bundle-assets://' + path
 }
 
-function createFile(path:string, data:string, encoding: 'base64' | 'ascii' | 'utf8'): Promise<string> {
-  encoding = encoding || 'utf8'
-  if(encoding.toLowerCase() === 'ascii') {
-    if(Array.isArray(data))
-      return RNFetchBlob.createFileASCII(path, data)
-    else {
-      const err = new TypeError('`data` of ASCII file must be an array with 0..255 numbers')
-      err.code = 'EINVAL'
-      return Promise.reject(err)
-    }
+function createFile(path: string, data: string, encoding: 'base64' | 'ascii' | 'utf8' = 'utf8'): Promise<string> {
+  if (encoding.toLowerCase() === 'ascii') {
+    return Array.isArray(data) ?
+      RNFetchBlob.createFileASCII(path, data) :
+      Promise.reject(addCode('EINVAL', new TypeError('`data` of ASCII file must be an array with 0..255 numbers')))
   }
   else {
     return RNFetchBlob.createFile(path, data, encoding)
@@ -82,17 +74,16 @@ function createFile(path:string, data:string, encoding: 'base64' | 'ascii' | 'ut
  * @return {Promise<RNFetchBlobWriteStream>} A promise resolves a `WriteStream` object.
  */
 function writeStream(
-  path : string,
-  encoding : 'utf8' | 'ascii' | 'base64',
-  append? : ?boolean,
-):Promise<RNFetchBlobWriteStream> {
-  if(!path)
-    throw Error('RNFetchBlob could not open file stream with empty `path`')
-  encoding = encoding || 'utf8'
-  append = append || false
+  path: string,
+  encoding?: 'utf8' | 'ascii' | 'base64' = 'base64',
+  append?: boolean = false,
+): Promise<RNFetchBlobWriteStream> {
+  if (typeof path !== 'string') {
+    return Promise.reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
+  }
   return new Promise((resolve, reject) => {
-    RNFetchBlob.writeStream(path, encoding || 'base64', append || false, (err, streamId:string) => {
-      if(err)
+    RNFetchBlob.writeStream(path, encoding, append, (err, streamId: string) => {
+      if (err)
         reject(new Error(err))
       else
         resolve(new RNFetchBlobWriteStream(streamId, encoding))
@@ -109,11 +100,14 @@ function writeStream(
  * @return {RNFetchBlobStream} RNFetchBlobStream stream instance.
  */
 function readStream(
-  path : string,
-  encoding : 'utf8' | 'ascii' | 'base64',
-  bufferSize? : ?number,
-  tick : ?number = 10
-):Promise<RNFetchBlobReadStream> {
+  path: string,
+  encoding: 'utf8' | 'ascii' | 'base64' = 'base64',
+  bufferSize?: number,
+  tick?: number = 10
+): Promise<RNFetchBlobReadStream> {
+  if (typeof path !== 'string') {
+    return Promise.reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
+  }
   return Promise.resolve(new RNFetchBlobReadStream(path, encoding, bufferSize, tick))
 }
 
@@ -122,8 +116,11 @@ function readStream(
  * @param  {string} path Path of directory to be created
  * @return {Promise}
  */
-function mkdir(path:string):Promise {
-    return RNFetchBlob.mkdir(path)
+function mkdir(path: string): Promise {
+  if (typeof path !== 'string') {
+    return Promise.reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
+  }
+  return RNFetchBlob.mkdir(path)
 }
 
 /**
@@ -131,7 +128,7 @@ function mkdir(path:string):Promise {
  * @param  {string} groupName Name of app group
  * @return {Promise}
  */
-function pathForAppGroup(groupName:string):Promise {
+function pathForAppGroup(groupName: string): Promise {
   return RNFetchBlob.pathForAppGroup(groupName)
 }
 
@@ -141,12 +138,9 @@ function pathForAppGroup(groupName:string):Promise {
  * @param  {'base64' | 'utf8' | 'ascii'} encoding Encoding of read stream.
  * @return {Promise<Array<number> | string>}
  */
-function readFile(path:string, encoding:string):Promise<any> {
-  encoding = encoding || 'utf8'
-  if(typeof path !== 'string') {
-    const err = new TypeError('Missing argument "path" ')
-    err.code = 'EINVAL'
-    return Promise.reject(err)
+function readFile(path: string, encoding: string = 'utf8'): Promise<any> {
+  if (typeof path !== 'string') {
+    return Promise.reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
   }
   return RNFetchBlob.readFile(path, encoding)
 }
@@ -158,50 +152,39 @@ function readFile(path:string, encoding:string):Promise<any> {
  * @param  {string} encoding Encoding of data (Optional).
  * @return {Promise}
  */
-function writeFile(path:string, data:string | Array<number>, encoding:?string):Promise {
-  encoding = encoding || 'utf8'
-  if(typeof path !== 'string') {
-    const err = new TypeError('Missing argument "path" ')
-    err.code = 'EINVAL'
-    return Promise.reject(err)
+function writeFile(path: string, data: string | Array<number>, encoding: ?string = 'utf8'): Promise {
+  if (typeof path !== 'string') {
+    return Promise.reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
   }
-  if(encoding.toLocaleLowerCase() === 'ascii') {
-    if(!Array.isArray(data)) {
-      const err = new TypeError('"data" must be an Array when encoding is "ascii"')
-      err.code = 'EINVAL'
-      return Promise.reject(err)
+  if (encoding.toLocaleLowerCase() === 'ascii') {
+    if (!Array.isArray(data)) {
+      return Promise.reject(addCode('EINVAL', new TypeError('"data" must be an Array when encoding is "ascii"')))
     }
     else
       return RNFetchBlob.writeFileArray(path, data, false)
   }
   else {
-    if(typeof data !== 'string') {
-      const err = new TypeError(`"data" must be a String when encoding is "utf8" or "base64", but it is "${typeof data}"`)
-      err.code = 'EINVAL'
-      return Promise.reject(err)
+    if (typeof data !== 'string') {
+      return Promise.reject(addCode('EINVAL', new TypeError(`"data" must be a String when encoding is "utf8" or "base64", but it is "${typeof data}"`)))
     }
     else
       return RNFetchBlob.writeFile(path, encoding, data, false)
   }
 }
 
-function appendFile(path:string, data:string | Array<number>, encoding:?string): Promise<number> {
-  encoding = encoding || 'utf8'
-  if(typeof path !== 'string') {
-    const err = new TypeError('Missing argument "path" ')
-    err.code = 'EINVAL'
-    return Promise.reject(err)
+function appendFile(path: string, data: string | Array<number>, encoding?: string = 'utf8'): Promise<number> {
+  if (typeof path !== 'string') {
+    return Promise.reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
   }
-  if(encoding.toLocaleLowerCase() === 'ascii') {
-    if(!Array.isArray(data)) {
-      const err = new TypeError('`data` of ASCII file must be an array with 0..255 numbers')
-      err.code = 'EINVAL'
-      return Promise.reject(err)
+  if (encoding.toLocaleLowerCase() === 'ascii') {
+    if (!Array.isArray(data)) {
+      return Promise.reject(addCode('EINVAL', new TypeError('`data` of ASCII file must be an array with 0..255 numbers')))
     }
     else
       return RNFetchBlob.writeFileArray(path, data, true)
-  } else {
-    if(typeof data !== 'string') {
+  }
+  else {
+    if (typeof data !== 'string') {
       const err = new TypeError(`"data" must be a String when encoding is "utf8" or "base64", but it is "${typeof data}"`)
       err.code = 'EINVAL'
       return Promise.reject(err)
@@ -216,18 +199,16 @@ function appendFile(path:string, data:string | Array<number>, encoding:?string):
  * @param  {string} path Target path
  * @return {RNFetchBlobFile}
  */
-function stat(path:string):Promise<RNFetchBlobFile> {
+function stat(path: string): Promise<RNFetchBlobFile> {
   return new Promise((resolve, reject) => {
-    if(typeof path !== 'string') {
-      const err = new TypeError('Missing argument "path" ')
-      err.code = 'EINVAL'
-      return reject(err)
+    if (typeof path !== 'string') {
+      return reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
     }
     RNFetchBlob.stat(path, (err, stat) => {
-      if(err)
+      if (err)
         reject(new Error(err))
       else {
-        if(stat) {
+        if (stat) {
           stat.size = parseInt(stat.size)
           stat.lastModified = parseInt(stat.lastModified)
         }
@@ -242,11 +223,14 @@ function stat(path:string):Promise<RNFetchBlobFile> {
  * @param  {Array<Object<string, string>>} pairs Array contains Key value pairs with key `path` and `mime`.
  * @return {Promise}
  */
-function scanFile(pairs:any):Promise {
+function scanFile(pairs: any): Promise {
   return new Promise((resolve, reject) => {
+    if (pairs === undefined) {
+      return reject(addCode('EINVAL', new TypeError('Missing argument')))
+    }
     RNFetchBlob.scanFile(pairs, (err) => {
-      if(err)
-        reject(new Error(err))
+      if (err)
+        reject(addCode('EUNSPECIFIED', new Error(err)))
       else
         resolve()
     })
@@ -254,52 +238,57 @@ function scanFile(pairs:any):Promise {
 }
 
 function hash(path: string, algorithm: string): Promise<string> {
-  if(typeof path !== 'string' || typeof algorithm !== 'string') {
-    const err = new TypeError('Missing argument "path" and/or "algorithm"')
-    err.code = 'EINVAL'
-    return Promise.reject(err)
+  if (typeof path !== 'string' || typeof algorithm !== 'string') {
+    return Promise.reject(addCode('EINVAL', new TypeError('Missing argument "path" and/or "algorithm"')))
   }
   return RNFetchBlob.hash(path, algorithm)
 }
 
-function cp(path:string, dest:string):Promise<boolean> {
+function cp(path: string, dest: string): Promise<boolean> {
   return new Promise((resolve, reject) => {
+    if (typeof path !== 'string' || typeof dest !== 'string') {
+      return reject(addCode('EINVAL', new TypeError('Missing argument "path" and/or "destination"')))
+    }
     RNFetchBlob.cp(path, dest, (err, res) => {
-      if(err)
-        reject(new Error(err))
+      if (err)
+        reject(addCode('EUNSPECIFIED', new Error(err)))
       else
         resolve(res)
     })
   })
 }
 
-function mv(path:string, dest:string):Promise<boolean> {
+function mv(path: string, dest: string): Promise<boolean> {
   return new Promise((resolve, reject) => {
+    if (typeof path !== 'string' || typeof dest !== 'string') {
+      return reject(addCode('EINVAL', new TypeError('Missing argument "path" and/or "destination"')))
+    }
     RNFetchBlob.mv(path, dest, (err, res) => {
-      if(err)
-        reject(new Error(err))
+      if (err)
+        reject(addCode('EUNSPECIFIED', new Error(err)))
       else
         resolve(res)
     })
   })
 }
 
-function lstat(path:string):Promise<Array<RNFetchBlobFile>> {
+function lstat(path: string): Promise<Array<RNFetchBlobFile>> {
   return new Promise((resolve, reject) => {
+    if (typeof path !== 'string') {
+      return reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
+    }
     RNFetchBlob.lstat(path, (err, stat) => {
-      if(err)
-        reject(new Error(err))
+      if (err)
+        reject(addCode('EUNSPECIFIED', new Error(err)))
       else
         resolve(stat)
     })
   })
 }
 
-function ls(path:string):Promise<Array<String>> {
-  if(typeof path !== 'string') {
-    const err = new TypeError('Missing argument "path" ')
-    err.code = 'EINVAL'
-    return Promise.reject(err)
+function ls(path: string): Promise<Array<String>> {
+  if (typeof path !== 'string') {
+    return Promise.reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
   }
   return RNFetchBlob.ls(path)
 }
@@ -309,11 +298,14 @@ function ls(path:string):Promise<Array<String>> {
  * @param  {string}   path:string Path of target file.
  * @return {Promise}
  */
-function unlink(path:string):Promise {
+function unlink(path: string): Promise {
   return new Promise((resolve, reject) => {
+    if (typeof path !== 'string') {
+      return reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
+    }
     RNFetchBlob.unlink(path, (err) => {
-      if(err) {
-        reject(new Error(err))
+      if (err) {
+        reject(addCode('EUNSPECIFIED', new Error(err)))
       }
       else
         resolve()
@@ -326,59 +318,70 @@ function unlink(path:string):Promise {
  * @param  {string} path Path to check
  * @return {Promise<boolean>}
  */
-function exists(path:string):Promise<boolean> {
+function exists(path: string): Promise<boolean> {
   return new Promise((resolve, reject) => {
+    if (typeof path !== 'string') {
+      return reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
+    }
     try {
       RNFetchBlob.exists(path, (exist) => {
         resolve(exist)
       })
-    } catch(err) {
-      reject(new Error(err))
+    }catch (err){
+      reject(addCode('EUNSPECIFIED', new Error(err)))
     }
   })
 
 }
 
-function slice(src:string, dest:string, start:number, end:number):Promise {
+function slice(src: string, dest: string, start: number, end: number): Promise {
+  if (typeof src !== 'string' || typeof dest !== 'string') {
+    return reject(addCode('EINVAL', new TypeError('Missing argument "src" and/or "destination"')))
+  }
+
   let p = Promise.resolve()
   let size = 0
+
   function normalize(num, size) {
-    if(num < 0)
+    if (num < 0)
       return Math.max(0, size + num)
-    if(!num && num !== 0)
+    if (!num && num !== 0)
       return size
     return num
   }
-  if(start < 0 || end < 0 || !start || !end) {
+
+  if (start < 0 || end < 0 || !start || !end) {
     p = p.then(() => stat(src))
-         .then((stat) => {
-           size = Math.floor(stat.size)
-           start = normalize(start || 0, size)
-           end = normalize(end, size)
-           return Promise.resolve()
-         })
+      .then((stat) => {
+        size = Math.floor(stat.size)
+        start = normalize(start || 0, size)
+        end = normalize(end, size)
+      })
   }
   return p.then(() => RNFetchBlob.slice(src, dest, start, end))
 }
 
-function isDir(path:string):Promise<bool> {
+function isDir(path: string): Promise<bool> {
   return new Promise((resolve, reject) => {
+    if (typeof path !== 'string') {
+      return reject(addCode('EINVAL', new TypeError('Missing argument "path" ')))
+    }
     try {
       RNFetchBlob.exists(path, (exist, isDir) => {
         resolve(isDir)
       })
-    } catch(err) {
-      reject(new Error(err))
+    }catch (err){
+      reject(addCode('EUNSPECIFIED', new Error(err)))
     }
   })
 
 }
 
-function df():Promise<{ free : number, total : number }> {
+function df(): Promise<{ free: number, total: number }> {
   return new Promise((resolve, reject) => {
     RNFetchBlob.df((err, stat) => {
-      if(err)
-        reject(new Error(err))
+      if (err)
+        reject(addCode('EUNSPECIFIED', new Error(err)))
       else
         resolve(stat)
     })
